new version of pascals script - renamed to domainviz.py
