+++ website-glen/api/api/domainviz.py
@@ -1,6 +1,6 @@
 ########################################################################################################################
 #                                                                                                                      #
-#   Author: Pascal Schläpfer, ETH Zürich, December 7th 2020                                                            #
+#   Author: Pascal Schläpfer, ETH Zürich, December 19th 2020                                                           #
 #   See below for function description                                                                                 #
 #                                                                                                                      #
 #   Acknowledgements:                                                                                                  #
@@ -68,13 +68,6 @@ except ImportError:
     print('Error, module statistics is required.')
     sys.exit()
 
-# test if json is installed in python
-try:
-    import json
-except ImportError:
-    print('Error, module json is required.')
-    sys.exit()
-
 # test if numpy is installed in python
 try:
     import numpy as np
@@ -130,6 +123,20 @@ except ImportError:
     print('Error, module Bio is required.')
     sys.exit()
 
+# test if Bio is installed in python
+try:
+    from Bio import ExPASy
+except ImportError:
+    print('Error, module Bio is required.')
+    sys.exit()
+
+# test if Bio is installed in python
+try:
+    from Bio.ExPASy import Prosite
+except ImportError:
+    print('Error, module Bio is required.')
+    sys.exit()
+
 # test if urllib is installed in python
 try:
     import urllib
@@ -158,7 +165,6 @@ except ImportError:
     print('Error, module time is required.')
     sys.exit()
 
-
 # test if datetime is installed in python
 try:
     from datetime import date
@@ -170,7 +176,7 @@ except ImportError:
 ########################################################################################################################
 #                                                                                                                      #
 #  f_process_results                                                                                                   #
-#  Main function of propplot.py. See help for further information                                                      #
+#  Main function of domainviz.py. See help for further information                                                     #
 #                                                                                                                      #
 #  Mandatory arguments:                                                                                                #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
@@ -196,10 +202,11 @@ except ImportError:
 #    - vcustom_scaling_on [string] gets converted into boolean: Allows for custom scaling of the figure. The standard  #
 #                                                               case is that this is not on, and thus the scaling is   #
 #                                                               the same regardless of the size of the protein. If     #
-#                                                               this is on, one can set sbp (default 100 bp per inch)  #
-#                                                               to scale the width of the figure.                      #
-#    - vscalingfigure [string] gets converted into float: Indicates the number of bp that are displayed per inch of    #
-#                                                         x-axis.                                                      #
+#                                                               this is on, one can set with option api (default 100   #
+#                                                               amino acids per inch) the number of amino acids        #
+#                                                               displayed per inch to scale the width of the figure.   #
+#    - vscalingfigure [string] gets converted into float: Indicates the number of amino acids per inch that are        #
+#                                                         displayed per inch of x-axis.                                #
 #    - vabsolute [string] gets converted into True/False: Indicates whether absolute numbers are displayed on the      #
 #                                                         y-axis.                                                      #
 #    - vwarnings [string] gets converted into True/False: Indicates whether warnings are written out.                  #
@@ -221,8 +228,8 @@ except ImportError:
 #                                              structure.                                                              #
 #    - proteingroup_pfam_colors.txt [file]: Same file as above, but for pfam plot.                                     #
 #    - proteingroup_combined_colors.txt [file]: Same file as above, but for combined plot.                             #
-#    - jobid_proteingroup_prosite.csv [file]: Results of Prosite for protein group.                                    #
-#    - jobid_proteingroup_pfam.csv [file]: Results of PFAM for protein group.                                          #
+#    - jobid_proteingroup_prosite_domain_results_per_aa.tsv [file]: Results of Prosite for protein group.              #
+#    - jobid_proteingroup_pfam_domain_results_per_aa.tsv [file]: Results of PFAM for protein group.                    #
 #    - jobid_prosite_res.tsv [file]: Results of Prosite for this job.                                                  #
 #    - jobid_pfam_res.tsv [file]: Results of PFAM for this job.                                                        #
 #    - Prosite_db_[first five amino acids of sequence] [file]: TSV database files for sequences run through Prosite.   #
@@ -231,9 +238,10 @@ except ImportError:
 ########################################################################################################################
 
 
-def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroupfile, vcolorfile, vignoredomainfile,
-                   vcutoff, vmaxcutoff, vcustom_scaling_on, vscalingfigure, vabsolute, vwarnings, vfrom_scratch,
-                   vnotolderthan):
+# noinspection PyTypeChecker
+def f_run_domainviz(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroupfile, vcolorfile, vignoredomainfile,
+                    vcutoff, vmaxcutoff, vcustom_scaling_on, vscalingfigure, vabsolute, vwarnings, vfrom_scratch,
+                    vnotolderthan):
     # Write initial cookie
     f_write_cookie(0, vsavefolder, vjobid, 'Job initialized')
 
@@ -245,12 +253,12 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
     vpaddingheight = 0.7
     vfontsize = 10
     vcurrentdate = int(re.sub('-', '', str(date.fromtimestamp(time.time()))))
-    
+
     # Write initial job:
-    f_write_log(vsavefolder, vjobid, 'Job initialized with following paramters\nJobid: ' + vjobid + '\nInputfile: ' +
+    f_write_log(vsavefolder, vjobid, 'Job initialized with following parameters\nJobid: ' + vjobid + '\nInputfile: ' +
                 vinputfile + '\nIgnoredb: ' + vignoredb + '\nSavefolder: ' + vsavefolder + '\nDbfolder: ' + vdbfolder +
                 '\nGroupfile: ' + vgroupfile + '\nColorfile: ' + vcolorfile + '\nIgnoredomainfile: ' +
-                vignoredomainfile + '\nCutoff: ' + vcutoff + '\nMaxcutoff: ' + vmaxcutoff + '\nCostomscaling: ' +
+                vignoredomainfile + '\nCutoff: ' + vcutoff + '\nMaxcutoff: ' + vmaxcutoff + '\nCustomscaling: ' +
                 vcustom_scaling_on + '\nScalingfigure: ' + vscalingfigure + '\nAbsolute: ' + vabsolute +
                 '\nWarnings: ' + vwarnings + '\nFrom_scratch: ' + vfrom_scratch + '\nNotolderthan: ' + vnotolderthan +
                 '\n\n', 'w')
@@ -298,7 +306,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
     vlen_dbid = 5  # Length of db names (defines how many different db files are made.)
     f_success(vsavefolder, vjobid)
 
-    # Read in headers and sequences that were used to produce the prosite and or pfam results
+    # Read in headers and sequences that were used to produce the Prosite and or PFAM results
     f_write_log(vsavefolder, vjobid, 'Reading in fasta file:\n', 'a')
     vheaders, vsequences = f_read_in_file(vinputfile, vsavefolder, vjobid)
     if vwarnings:
@@ -345,6 +353,8 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
 
         # Process every header
         for vheader_id, vheader in enumerate(vheaders):
+            if vwarnings:
+                print('Searching header ' + str(vheader_id + 1) + ' of ' + str(len(vheaders)) + '.')
             f_write_log(vsavefolder, vjobid, 'Processing header: ' + vheader + '\n', 'a')
             vfound = False  # Defines if the header was found in the db
             vdbid = vsequences[vheader_id][0:vlen_dbid]  # Defines the db file that should be searched.
@@ -362,7 +372,8 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                                 vfound = True  # if it is, say that the sequence was found in the db.
                                 f_write_log(vsavefolder, vjobid, 'Found entry in DB ' + vdbid + '\n', 'a')
                                 f_write_pfam_prosite_res(vprositefile, vheader, vsplitentry[1:], False, True,
-                                                         vsavefolder, vjobid)  # Write the record to the output file.
+                                                         vsavefolder, vjobid)  # Write the record to the
+                                # output file.
                     vfh_db_prosite.close()
                 except IOError:  # If the db can not be read, it likely does not exist.
                     f_write_log(vsavefolder, vjobid, 'does not yet exist\n', 'a')
@@ -371,17 +382,18 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
 
             if not vfound:  # If the sequence has not been found in a db.
                 f_write_log(vsavefolder, vjobid, 'Sending sequence to Prosite\n', 'a')
-                ventry_found = f_run_sequences_through_prosite(vheader, vsequences[vheader_id], vsavefolder, vjobid,
-                                                               vwarnings)
-                # Search for the sequence in prosite.
-                f_write_pfam_prosite_db(vdbfiles, vdbid, ventry_found, vsavefolder, vjobid, vcurrentdate)  # Write the
-                # results into the db.
+                ventry_found = f_run_sequences_through_prosite(vdbfolder, vheader, vsequences[vheader_id], vsavefolder,
+                                                               vjobid, vwarnings, vcurrentdate)  # Search for the
+                # sequence in Prosite.
+                f_write_pfam_prosite_db(vdbfiles, vdbid, ventry_found, vsavefolder, vjobid, vcurrentdate)  # Write
+                # the results into the db.
                 f_write_pfam_prosite_res(vprositefile, vheader, ventry_found, False, False, vsavefolder, vjobid)
                 # Write the results into the results file.
             # Update cookies
             vcid = math.floor((vheader_id + 1) / len(vheaders) * 100 / 5) + 1
             if vcid > 1:
-                f_write_cookie(vcid, vsavefolder, vjobid, '')
+                f_write_cookie(vcid, vsavefolder, vjobid, str((vcid - 1) * 5) + '% of sequences evaluated for ' +
+                               'prosite domains.')
 
     vprositedata = f_read_tsv(vprositefile, vsavefolder, vjobid)  # Read the data of prosite from the result file.
     f_write_log(vsavefolder, vjobid, 'Successfully gathered Prosite domain results.\n', 'a')
@@ -412,12 +424,6 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
     if not vpfam_already_done:
         f_write_log(vsavefolder, vjobid, 'Create new PFAM output file: ', 'a')
         vfh_pfam_output = open(vpfamfile, 'w')  # writing the header of the pfam result file.
-        # vfh_pfam_output.write('Sequence id\tFamily id\tFamily Accession\tClan\tEnv. Start\tEnv. End\tAli. Start\t'
-        #                      'Ali. End\tModel Start\tModel End\tBit Score\tInd. E-value\tCond. E-value\tDescription\t'
-        #                      'aliIdCount\taliL\taliM\taliN\taliSim\taliSimCount\taliaseq\talicsline\talihindex\t'
-        #                      'alimline\talimmline\talimodel\talintseq\talippline\talirfline\talisqacc\talisqdesc\t'
-        #                      'alisqfrom\talisqname\talisqto\tbias\tdisplay\tis_included\tis_reported\toasc\t'
-        #                      'outcompeted\tsignificant\tuniq\n')
         vfh_pfam_output.write('Sequence id\tFamily id\tFamily Accession\tClan\tEnv. Start\tEnv. End\tAli. Start\t'
                               'Ali. End\tModel Start\tModel End\tBit Score\tInd. E-value\tCond. E-value\tDescription\t'
                               'outcompeted\tsignificant\tuniq\n')
@@ -432,6 +438,8 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
 
         # Process every header
         for vheader_id, vheader in enumerate(vheaders):
+            if vwarnings:
+                print('Searching header ' + str(vheader_id + 1) + ' of ' + str(len(vheaders)) + '.')
             f_write_log(vsavefolder, vjobid, 'Processing header: ' + vheader + '\n', 'a')
             vfound = False
             vdbid = vsequences[vheader_id][0:vlen_dbid]
@@ -474,8 +482,9 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             # Update cookies
             vcid = math.floor((vheader_id + 1) / len(vheaders) * 100 / 5) + 22
             if vcid > 22:
-                f_write_cookie(vcid, vsavefolder, vjobid, '')
-                
+                f_write_cookie(vcid, vsavefolder, vjobid, str((vcid - 22) * 5) + '% of sequences evaluated for pfam ' +
+                               'domains.')
+
     vpfamdata = f_read_tsv(vpfamfile, vsavefolder, vjobid)
     f_write_log(vsavefolder, vjobid, 'Successfully gathered PFAM domain results.\n', 'a')
     f_write_cookie(43, vsavefolder, vjobid, 'Finished searching PFAM')
@@ -595,7 +604,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
     f_write_cookie(60, vsavefolder, vjobid, 'Finished reading and processing additional data')
 
     # Per group of protein: get annotations of Prosite and PFAM domains and make one plot per protein group.
-    f_write_log(vsavefolder, vjobid, 'Collecting Prostie and PFAM domain information per protein group\n', 'a')
+    f_write_log(vsavefolder, vjobid, 'Collecting Prosite and PFAM domain information per protein group\n', 'a')
     vprositedomainsofgroup_all = []
     vprositedomainsingenes_all = []
     vpfamdomainsofgroup_all = []
@@ -670,13 +679,13 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                             for vp, vpd in enumerate(vprositedomains_u):  # Go through all domains
                                 if vprositedata[vheader_id][3] == vpd:  # Check if the domain is the domain we currently
                                     # look at
-                                    vmedianstartbp = f_float2int((int(vprositedata[vheader_id][1]) - 1) /
+                                    vmedianstartaa = f_float2int((int(vprositedata[vheader_id][1]) - 1) /
                                                                  len(vsequences[vh]) * vmedlengroup[vug])  # Calculate
                                     # the start position relative to the median length of the protein group:
-                                    vmedianendbp = f_float2int((int(vprositedata[vheader_id][2]) - 1) /
+                                    vmedianendaa = f_float2int((int(vprositedata[vheader_id][2]) - 1) /
                                                                len(vsequences[vh]) * vmedlengroup[vug])  # Calculate the
                                     # end position relative to the median length of the protein group:
-                                    vprositedomainsofgroup[vp, vmedianstartbp:vmedianendbp] += 1  # Mark the placing of
+                                    vprositedomainsofgroup[vp, vmedianstartaa:vmedianendaa] += 1  # Mark the placing of
                                     # the domain
                                     vprositedomainsingenes[vp, vh] = 1  # Define that the domain has been found for that
                                     # gene
@@ -688,13 +697,13 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                             # currently look at
                             for vp, vpd in enumerate(vpfamdomains_u):  # Go through all domains
                                 if vpfamdata[vheader_id][1] == vpd:  # If the domain is the domain we currently look at
-                                    vmedianstartbp = f_float2int((int(vpfamdata[vheader_id][6]) - 1) /
+                                    vmedianstartaa = f_float2int((int(vpfamdata[vheader_id][6]) - 1) /
                                                                  len(vsequences[vh]) * vmedlengroup[vug])  # Calculate
                                     # the start position relative to the median length of the protein group:
-                                    vmedianendbp = f_float2int((int(vpfamdata[vheader_id][7]) - 1) /
+                                    vmedianendaa = f_float2int((int(vpfamdata[vheader_id][7]) - 1) /
                                                                len(vsequences[vh]) * vmedlengroup[vug])  # Calculate the
                                     # end position relative to the median length of the protein group:
-                                    vpfamdomainsofgroup[vp, vmedianstartbp:vmedianendbp] += 1  # Mark the placing of the
+                                    vpfamdomainsofgroup[vp, vmedianstartaa:vmedianendaa] += 1  # Mark the placing of the
                                     # domain
                                     vpfamdomainsingenes[vp, vh] = 1  # Define that the domain has been found for
                                     # that gene
@@ -723,13 +732,14 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             for vprd in range(vsize_of_prosite_data[0]):
                 for vj in range(vsize_of_prosite_data[1]):
                     if vprositedomainsofgroup[vprd][vj] != 0:
-                        vprositedomainsofgroup_rel[vprd][vj] = vprositedomainsofgroup[vprd][vj]/vn_prot_per_group * 100
+                        vprositedomainsofgroup_rel[vprd][vj] = vprositedomainsofgroup[vprd][vj] / vn_prot_per_group * \
+                                                               100
         if vpfamfile != '':
             vpfamdomainsofgroup_rel = np.zeros((len(vpfamdomains_u), vmedlengroup[vug]), dtype=float)
             for vpfd in range(vsize_of_pfam_data[0]):
                 for vj in range(vsize_of_pfam_data[1]):
                     if vpfamdomainsofgroup[vpfd][vj] != 0:
-                        vpfamdomainsofgroup_rel[vpfd][vj] = vpfamdomainsofgroup[vpfd][vj]/vn_prot_per_group * 100
+                        vpfamdomainsofgroup_rel[vpfd][vj] = vpfamdomainsofgroup[vpfd][vj] / vn_prot_per_group * 100
         vprositedomainsofgroup_rel_all.append(vprositedomainsofgroup_rel)
         vpfamdomainsofgroup_rel_all.append(vpfamdomainsofgroup_rel)
         vn_prot_per_group_all.append(vn_prot_per_group)
@@ -748,23 +758,52 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
         vpfamdomainsofgroup_all.append(vpfamdomainsofgroup)
 
         # Save of data
-        f_write_log(vsavefolder, vjobid, 'Save domain data of group as .csvs: ', 'a')
+        f_write_log(vsavefolder, vjobid, 'Save domain data of group as .tsvs: ', 'a')
         if vprositefile != '':
             if vsavefolder != '':
-                np.savetxt(join(vsavefolder, vjobid + '_' + vgitem + '_prosite.csv'), vprositedomainsofgroup,
-                           delimiter="\t")
+                vfilename = join(vsavefolder, vjobid + '_' + vgitem + '_prosite_domain_results_per_aa.tsv')
             else:
-                np.savetxt(vjobid + '_' + vgitem + '_prosite.csv', vprositedomainsofgroup, delimiter="\t")
+                vfilename = vjobid + '_' + vgitem + '_prosite_domain_results_per_aa.tsv'
+            vfh_filename = open(vfilename, 'w')
+
+            # Write header:
+            vfh_filename.write('Median protein (' + vgitem + ') amino acid position')
+            for item in vprositedomains_u:
+                vfh_filename.write('\t' + item)
+            vfh_filename.write('\n')
+
+            # Write content
+            for vj in range(vsize_of_prosite_data[1]):
+                vfh_filename.write(str(vj + 1))
+                for vprd in range(vsize_of_prosite_data[0]):
+                    vfh_filename.write('\t' + str(vprositedomainsofgroup[vprd][vj]))
+                vfh_filename.write('\n')
+            vfh_filename.close()
         if vpfamfile != '':
             if vsavefolder != '':
-                np.savetxt(join(vsavefolder, vjobid + '_' + vgitem + '_pfam.csv'), vpfamdomainsofgroup, delimiter="\t")
+                vfilename = join(vsavefolder, vjobid + '_' + vgitem + '_pfam_domain_results_per_aa.tsv')
             else:
-                np.savetxt(vjobid + '_' + vgitem + '_pfam.csv', vpfamdomainsofgroup, delimiter="\t")
+                vfilename = vjobid + '_' + vgitem + '_pfam_domain_results_per_aa.tsv'
+            vfh_filename = open(vfilename, 'w')
+
+            # Write header:
+            vfh_filename.write('Median protein (' + vgitem + ') amino acid position')
+            for item in vpfamdomains_u:
+                vfh_filename.write('\t' + item)
+            vfh_filename.write('\n')
+
+            # Write content
+            for vj in range(vsize_of_pfam_data[1]):
+                vfh_filename.write(str(vj + 1))
+                for vpfd in range(vsize_of_pfam_data[0]):
+                    vfh_filename.write('\t' + str(vpfamdomainsofgroup[vpfd][vj]))
+                vfh_filename.write('\n')
+            vfh_filename.close()
         f_success(vsavefolder, vjobid)
 
         # Plot domains into figure with all domains of all groups to get coloring identical
         f_write_log(vsavefolder, vjobid, 'Add domains into dummy figure to get consistent domain data of group as'
-                                         ' .csvs: ', 'a')
+                                         ' .tsvs: ', 'a')
         vmaxes = []
         for vprd in range(vsize_of_prosite_data[0]):
             vmaxes.append(max(vprositedomainsofgroup_rel[vprd]))
@@ -774,7 +813,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
         bin_edges = np.arange(vsize_of_prosite_data[1] + 1)
         for vmaxi in vmaxids:
             if vmaxi < vsize_of_prosite_data[0]:
-                if max(vprositedomainsofgroup[vmaxi]) > (vmaxcutoff * 100) and float(
+                if max(vprositedomainsofgroup_rel[vmaxi]) > (vmaxcutoff * 100) and float(
                         sum(vprositedomainsingenes[vmaxi])) / float(vn_prot_per_group) > vcutoff and \
                         vprositedomains_u_ignore[vmaxi] == 0 and vprositedomains_u_process[vmaxi]:
                     if vprositedomains_u_color[vmaxi] == '':
@@ -790,7 +829,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                     vprositedomains_u_process[vmaxi] = False
             else:
                 vimod = vmaxi - vsize_of_prosite_data[0]
-                if max(vpfamdomainsofgroup[vimod]) > (vmaxcutoff * 100) and float(
+                if max(vpfamdomainsofgroup_rel[vimod]) > (vmaxcutoff * 100) and float(
                         sum(vpfamdomainsingenes[vimod])) / float(vn_prot_per_group) > vcutoff and \
                         vpfamdomains_u_ignore[vimod] == 0 and vpfamdomains_u_process[vimod]:
                     if vpfamdomains_u_color[vimod] == '':
@@ -836,8 +875,8 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             vmaxids = np.argsort(vmaxes)[::-1]
             bin_edges = np.arange(vsize_of_prosite_data[1] + 1)
             for vheader_id in vmaxids:
-                if max(vprositedomainsofgroup[vheader_id]) > (vmaxcutoff * 100) and \
-                        float(sum(vprositedomainsingenes[vheader_id]))/float(vn_prot_per_group) > vcutoff and \
+                if max(vprositedomainsofgroup_rel[vheader_id]) > (vmaxcutoff * 100) and \
+                        float(sum(vprositedomainsingenes[vheader_id])) / float(vn_prot_per_group) > vcutoff and \
                         vprositedomains_u_ignore[vheader_id] == 0:
                     # Create function
                     a = min(bin_edges[:-1])  # integral lower limit
@@ -864,10 +903,12 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                     # Plot Polygon
                     ax.add_patch(poly)
             plt.xlim(min(bin_edges), max(bin_edges))
-            if not vabsolute:
-                plt.ylim(0, 100)
+            if vabsolute:
+                plt.ylim(0, vn_prot_per_group)  # maximum of proteins in group
+            else:
+                plt.ylim(0, 100)  # 100 %
             plt.grid(axis='y', alpha=0.75)
-            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' bp', fontsize=vfontsize)
+            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' amino acids', fontsize=vfontsize)
             if vabsolute:
                 plt.ylabel('# of occurrences', fontsize=vfontsize)
             else:
@@ -899,7 +940,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             vmaxids = np.argsort(vmaxes)[::-1]
             bin_edges = np.arange(vsize_of_pfam_data[1] + 1)
             for vheader_id in vmaxids:
-                if max(vpfamdomainsofgroup[vheader_id]) > (vmaxcutoff * 100) and \
+                if max(vpfamdomainsofgroup_rel[vheader_id]) > (vmaxcutoff * 100) and \
                         float(sum(vpfamdomainsingenes[vheader_id])) / \
                         float(vn_prot_per_group) > vcutoff and vpfamdomains_u_ignore[vheader_id] == 0:
                     # Create function
@@ -928,10 +969,12 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                     ax.add_patch(poly)
 
             plt.xlim(min(bin_edges), max(bin_edges))
-            if not vabsolute:
+            if vabsolute:
+                plt.ylim(0, vn_prot_per_group)
+            else:
                 plt.ylim(0, 100)
             plt.grid(axis='y', alpha=0.75)
-            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' bp', fontsize=vfontsize)
+            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' amino acids', fontsize=vfontsize)
             if vabsolute:
                 plt.ylabel('# of occurrences', fontsize=vfontsize)
             else:
@@ -966,7 +1009,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             bin_edges = np.arange(vsize_of_prosite_data[1] + 1)
             for vmaxi in vmaxids:
                 if vmaxi < vsize_of_prosite_data[0]:
-                    if max(vprositedomainsofgroup[vmaxi]) > (vmaxcutoff * 100) and float(
+                    if max(vprositedomainsofgroup_rel[vmaxi]) > (vmaxcutoff * 100) and float(
                             sum(vprositedomainsingenes[vmaxi])) / float(vn_prot_per_group) > vcutoff and \
                             vprositedomains_u_ignore[vmaxi] == 0:
                         # Create function
@@ -995,7 +1038,7 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                         ax.add_patch(poly)
                 else:
                     vimod = vmaxi - vsize_of_prosite_data[0]
-                    if max(vpfamdomainsofgroup[vimod]) > (vmaxcutoff * 100) and float(
+                    if max(vpfamdomainsofgroup_rel[vimod]) > (vmaxcutoff * 100) and float(
                             sum(vpfamdomainsingenes[vimod])) / float(vn_prot_per_group) > vcutoff and \
                             vpfamdomains_u_ignore[vimod] == 0:
                         # Create function
@@ -1024,10 +1067,12 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
                         ax.add_patch(poly)
 
             plt.xlim(min(bin_edges), max(bin_edges))
-            if not vabsolute:
+            if vabsolute:
+                plt.ylim(0, vn_prot_per_group)
+            else:
                 plt.ylim(0, 100)
             plt.grid(axis='y', alpha=0.75)
-            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' bp', fontsize=vfontsize)
+            plt.xlabel('Median length: ' + str(vmedlengroup[vug]) + ' amino acids', fontsize=vfontsize)
             if vabsolute:
                 plt.ylabel('# of occurrences', fontsize=vfontsize)
             else:
@@ -1041,8 +1086,9 @@ def f_run_propplot(vjobid, vinputfile, vignoredb, vsavefolder, vdbfolder, vgroup
             else:
                 plt.savefig(vjobid + '_' + vgitem + '_combined.pdf')
     # Write fifth cookie
-    f_write_cookie(100, vsavefolder, vjobid, 'Job ' + vjobid + 'successful')
     f_write_log(vsavefolder, vjobid, 'Job ' + vjobid + ' ran successfully.\n', 'a')
+    time.sleep(5)  # make sure pdfs are created
+    f_write_cookie(100, vsavefolder, vjobid, 'Job ' + vjobid + 'successful')
     print('done')
 
 
@@ -1277,14 +1323,14 @@ def f_read_in_groupfile(vfile, vheaders, vsavefolder, vjobid):
                         vhit += 1
                     else:
                         print('Warning: group domain file has more than one entry for ' + vitem + '. Only first '
-                              'instance is used.')
+                                                                                                  'instance is used.')
                         f_write_log(vsavefolder, vjobid, 'Warning: group domain file has more than one entry '
                                                          'for ' + vitem + '. Only first instance is used.\n', 'a')
             vfile_fh.close()
             if vhit == 0:
                 f_write_log(vsavefolder, vjobid, 'Warning: no protein group entry found for ' + vitem +
-                                                 '. Please add one in ' + vfile + '\nUsing "ProteinGroup" as group name'
-                                                 ' instead.\n', 'a')
+                            '. Please add one in ' + vfile + '\nUsing "ProteinGroup" as group name'
+                                                             ' instead.\n', 'a')
                 print('Warning: no protein group entry found for ' + vitem + '. Please add one in ' + vfile)
                 vgroup.append('ProteinGroup')
                 vgroup_u.append('ProteinGroup')
@@ -1392,7 +1438,7 @@ def f_read_in_file(vfile, vsavefolder, vjobid):
         vheaders.append(vtempheader)
         vsequences.append(vtempsequence)
     f_write_log(vsavefolder, vjobid, 'Read ' + str(len(vheaders)) + ' headers and ' + str(len(vsequences)) +
-                                     ' sequences.\n', 'a')
+                ' sequences.\n', 'a')
     return vheaders, vsequences
 
 
@@ -1403,11 +1449,14 @@ def f_read_in_file(vfile, vsavefolder, vjobid):
 #  found.                                                                                                              #
 #                                                                                                                      #
 #  Mandatory arguments:                                                                                                #
+#    - vdbfolder [string]: Indicates the location of the folder where the databases of previous results should be      #
+#                          built up.                                                                                   #
 #    - vhead [string]: Fasta sequence header.                                                                          #
 #    - vseq [string]: Protein sequence.                                                                                #
 #    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
 #    - vwarnings_prosite [boolean]: If warnings should be given out or not.                                            #
+#    - vcurrentdate [string]: Indicates the currend date that the record is from.                                      #
 #                                                                                                                      #
 #  Output:                                                                                                             #
 #    - vout_all [list]: List of lists, containing the results as sublist, with the sequence header as the first item,  #
@@ -1416,7 +1465,7 @@ def f_read_in_file(vfile, vsavefolder, vjobid):
 ########################################################################################################################
 
 
-def f_run_sequences_through_prosite(vhead, vseq, vsavefolder, vjobid, vwarnings_prosite):
+def f_run_sequences_through_prosite(vdbfolder, vhead, vseq, vsavefolder, vjobid, vwarnings_prosite, vcurrentdate):
     # Write that sequences are run to Prosite to the stdout
     f_write_log(vsavefolder, vjobid, 'Submitted Prosite search for sequence "' + vseq[0:16] + '...".\n', 'a')
     if vwarnings_prosite:
@@ -1424,7 +1473,7 @@ def f_run_sequences_through_prosite(vhead, vseq, vsavefolder, vjobid, vwarnings_
 
     # Check for sequence length
     vout_all = []
-    if len(vseq) >= 16:
+    if 16 <= len(vseq):
         # Run sequence
         try:
             # If a 404 error occurs, then the reason might be that the mirror https://prosite.expasy.org changed to
@@ -1470,7 +1519,82 @@ def f_run_sequences_through_prosite(vhead, vseq, vsavefolder, vjobid, vwarnings_
         vout_all.append(vout)
         if vwarnings_prosite:
             print(vhead + ' is too short to be run through Prosite (<16 amino acids).')
-    return vout_all
+
+    # Get domain information using the domain ids
+    vout_all2 = f_get_prosite_domain_info(vout_all, vdbfolder, vcurrentdate)
+    return vout_all2
+
+
+########################################################################################################################
+#                                                                                                                      #
+#  f_get_prosite_domain_info                                                                                           #
+#  Runs a the domain identifier of prosite first through the internal databases, and if there is no record, goes to    #
+#  Prosite to find the description of the domain info.                                                                 #
+#                                                                                                                      #
+#  Mandatory arguments:                                                                                                #
+#    - vrecords [string]: Prosite domain results of a protein                                                          #
+#    - vdbfolder [string]: Indicates the location of the folder where the databases of previous results should be      #
+#                          built up.                                                                                   #
+#    - vcurrentdate [string]: Indicates the currend date that the record is from.                                      #
+#                                                                                                                      #
+#  Output:                                                                                                             #
+#    - vrecords_out [list]: List of lists, containing the results as sublist, with the sequence header as the first    #
+#                           item, followed by all other information collected from Prosite. the domain ID is swapped   #
+#                           with the description of the domain, and the domain id is listed last                       #
+#                                                                                                                      #
+########################################################################################################################
+
+
+def f_get_prosite_domain_info(vrecords, vdbfolder, vcurrentdate):
+    vprosite_domain_db = join(vdbfolder, 'Prosite_domain_db')
+    vrecords_out = []
+    vinit_db = False
+    for vrecord in vrecords:
+        if vrecord[3] != '':
+            print(vrecord)
+            print(vrecord[3])
+            vfound = False
+            try:
+                vfh_pdb = open(vprosite_domain_db, 'r')
+                for line in vfh_pdb:
+                    vsplit = line.rstrip('\n').split('\t')
+                    if vsplit[1] == vrecord[3]:
+                        vnew_record = []
+                        for vrecord_i, item in enumerate(vrecord):
+                            if vrecord_i != 3 and vrecord_i != 7:
+                                vnew_record.append(item)
+                            if vrecord_i == 3:
+                                vnew_record.append(vsplit[2])
+                            if vrecord_i == 7:
+                                vnew_record.append(vrecord[3])
+                        vrecords_out.append(vnew_record)
+                        vfound = True
+                        break
+                vfh_pdb.close()
+            except IOError:
+                vinit_db = True
+            if not vfound:
+                handle = ExPASy.get_prosite_raw(vrecord[3])
+                record = Prosite.read(handle)
+                handle.close()
+                vnew_record = []
+                for vrecord_i, item in enumerate(vrecord):
+                    if vrecord_i != 3 and vrecord_i != 7:
+                        vnew_record.append(item)
+                    if vrecord_i == 3:
+                        vnew_record.append(record.name)
+                    if vrecord_i == 7:
+                        vnew_record.append(vrecord[3])
+                vrecords_out.append(vnew_record)
+                if vinit_db:
+                    vfh_pdb = open(vprosite_domain_db, 'w')
+                else:
+                    vfh_pdb = open(vprosite_domain_db, 'a')
+                vfh_pdb.write(str(vcurrentdate) + '\t' + vrecord[3] + '\t' + record.name + '\n')
+                vfh_pdb.close()
+        else:
+            vrecords_out.append(vrecord)
+    return vrecords_out
 
 
 ########################################################################################################################
@@ -1494,6 +1618,11 @@ def f_run_sequences_through_prosite(vhead, vseq, vsavefolder, vjobid, vwarnings_
 
 
 def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfam):
+    # Constants:
+    vmax_pfam_503_iterations = 6
+    vinitial_503_wait_time_multiplier = 5
+    vcurrent_503_try = 0
+
     # Write that sequences are run to PFAM to the stdout
     f_write_log(vsavefolder, vjobid, 'Submitted PFAM search for sequence "' + vseq[0:16] + '...".\n', 'a')
     if vwarnings_pfam:
@@ -1501,10 +1630,60 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
 
     # Check for sequence length
     vout_all = []
-    if len(vseq) >= 16:
+    if 16 <= len(vseq) <= 5000:
         # Run sequence
         try:
-            vheader, vresults = f_query_pfam(vseq, vsavefolder, vjobid, 0, 0)  # Call PFAM with sequence
+            verrorcode, vheader, vresults = f_query_pfam(vseq, vsavefolder, vjobid, vwarnings_pfam)  # Call PFAM with
+            # sequence
+            if verrorcode == 503 or verrorcode == 500 or verrorcode == -2:
+                verrorcode_2_display = str(verrorcode)
+                if vwarnings_pfam:
+                    if verrorcode_2_display == '-2':
+                        verrorcode_2_display = 'A seldom error that usually can be rescued was found. '
+                    else:
+                        verrorcode_2_display = 'HTTP ' + verrorcode_2_display + ' Error occurred. '
+                    print(verrorcode_2_display + 'Try to rescue.')
+                f_write_log(vsavefolder, vjobid, verrorcode_2_display + '\nTrying to rescue.\n', 'a')
+                vcurrent_503_try += 1
+                vpassed = False
+                while vcurrent_503_try <= vmax_pfam_503_iterations and vpassed is False:
+                    if vwarnings_pfam:
+                        print('HTTP 500, 503, or a seldom other Error occurred. Rescue attempt #' +
+                              str(vcurrent_503_try) + ' of ' + str(vmax_pfam_503_iterations))
+                    f_write_log(vsavefolder, vjobid,
+                                'Rescue attempt #' + str(vcurrent_503_try) + ' of ' + str(vmax_pfam_503_iterations) +
+                                '.\n', 'a')
+                    vwaittime = (4 ^ vcurrent_503_try) * vinitial_503_wait_time_multiplier
+                    print('Waiting ' + str(vwaittime) + ' seconds.')
+                    time.sleep(vwaittime)
+                    verrorcode, vheader, vresults = f_query_pfam(vseq, vsavefolder, vjobid, vwarnings_pfam)  # Call PFAM
+                    # with sequence
+                    if verrorcode == -1:
+                        vpassed = True
+                    elif verrorcode == 503:
+                        if vwarnings_pfam:
+                            print('HTTP 503 Error occurred again.')
+                    elif verrorcode == 500:
+                        if vwarnings_pfam:
+                            print('HTTP 500 Error occurred again.')
+                    elif verrorcode == -2:
+                        if vwarnings_pfam:
+                            print('Other error than HTTP occurred again. Experience says: just try another time.')
+                    else:
+                        f_handle_pfam_error(verrorcode, vsavefolder, vjobid, vwarnings_pfam)
+                    vcurrent_503_try += 1
+                if vpassed is False:
+                    f_write_log(vsavefolder, vjobid, 'Was unable to rescue 500/503.\nAbort.\n', 'a')
+                    print('HTTPError 500 or 503, or other seldom errors occurred too many times when trying to access '
+                          'PFAM result.')
+                    f_write_log(vsavefolder, vjobid, 'HTTPError 500, 503 or seldom other errors occurred when trying '
+                                                     'to access PFAM result.\n', 'a')
+                    f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError 500, 503 or seldom other errors occurred too '
+                                                            'many times when trying to access PFAM result.')
+                    sys.exit()
+            else:
+                f_handle_pfam_error(verrorcode, vsavefolder, vjobid, vwarnings_pfam)
+
             if len(vresults) > 0:  # If at least one result was found
                 for vr, vresult in enumerate(vresults):  # For every result (id vr)
                     vout = [vseq]  # Attach sequence as identifier
@@ -1513,7 +1692,7 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
                         vout.append('')
                     # for vi_41, vitem in enumerate(vresult):  # Check for all attributes and fill them into the output
                     for vi_16, vitem in enumerate(
-                                vresult):  # Check for all attributes and fill them into the output
+                            vresult):  # Check for all attributes and fill them into the output
                         # vout = f_dissect_pfam_key(vout, vheader[vi_41], vitem)
                         vout = f_dissect_pfam_key(vout, vheader[vi_16], vitem)
                     vout_all.append(vout)  # Append the output of this result to the output of all other results
@@ -1535,7 +1714,7 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
             vout_all.append(vout)
             if vwarnings_pfam:
                 print(vhead + ' has no results in pfam.')
-    else:
+    elif 16 > len(vseq):
         f_write_log(vsavefolder, vjobid, 'Warning, sequence too short (<16 amino acids).\n', 'a')
         vout = [vseq]
         # for _ in range(42 - 1):
@@ -1544,7 +1723,68 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
         vout_all.append(vout)
         if vwarnings_pfam:
             print(vhead + ' is too short to be run through PFAM (<16 amino acids).')
+    elif len(vseq) > 5000:
+        f_write_log(vsavefolder, vjobid, 'Warning, sequence too long (>5000 amino acids).\n', 'a')
+        vout = [vseq]
+        # for _ in range(42 - 1):
+        for _ in range(17 - 1):
+            vout.append('')
+        vout_all.append(vout)
+        if vwarnings_pfam:
+            print(vhead + ' is too long to be run through PFAM (>5000 amino acids).')
     return vout_all  # Return all results
+
+
+########################################################################################################################
+#                                                                                                                      #
+#  f_handle_pfam_error                                                                                                 #
+#  Handles errors given out by running sequences through PFAM                                                          #
+#                                                                                                                      #
+#  Mandatory arguments:                                                                                                #
+#    - verrorcode [string]: Error code that needs handling.                                                            #
+#    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
+#    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
+#    - vwarnings_pfam [boolean]: If warnings should be given out or not.                                               #
+#                                                                                                                      #
+#  Output:                                                                                                             #
+#    - vout_all [list]: List of lists, containing the results as sublist, with the sequence header as the first item,  #
+#                       followed by all other information collected from pfam.                                         #
+#                                                                                                                      #
+########################################################################################################################
+
+
+def f_handle_pfam_error(verrorcode, vsavefolder, vjobid, vwarnings_pfam):
+    if verrorcode == -1:
+        if vwarnings_pfam:
+            print('Results of search received.')
+    elif verrorcode == 400:
+        print('HTTPError 400 occurred when trying to access PFAM result.')
+        f_write_log(vsavefolder, vjobid, 'HTTPError 400 occurred when trying to access PFAM '
+                                         'result.\n', 'a')
+        f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError 400 occurred when trying to access '
+                                                'PFAM result.')
+        sys.exit()
+    elif verrorcode == 404:
+        print('HTTPError 404 occurred when trying to access PFAM result.')
+        f_write_log(vsavefolder, vjobid, 'HTTPError 404 occurred when trying to access PFAM '
+                                         'result.\n', 'a')
+        f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError 404 occurred when trying to access '
+                                                'PFAM result.')
+        sys.exit()
+    elif verrorcode == -2:
+        print('Error other than HTTPError occurred when trying to access PFAM result.')
+        f_write_log(vsavefolder, vjobid, 'Error other than HTTPError occurred when trying to access PFAM '
+                                         'result.\n', 'a')
+        f_write_cookie(-1, vsavefolder, vjobid, 'Error other than HTTPError occurred when trying to access '
+                                                'PFAM result.')
+        sys.exit()
+    else:
+        print('Unknown error ' + str(verrorcode) + ' occurred: ')
+        f_write_log(vsavefolder, vjobid, 'Unknown error ' + str(verrorcode) + ' occurred while trying '
+                                                                              'to reach PFAM.\n', 'a')
+        f_write_cookie(-1, vsavefolder, vjobid, 'Unknown error ' + str(verrorcode) + ' occurred while trying '
+                                                                                     'to reach PFAM.')
+        sys.exit()
 
 
 ########################################################################################################################
@@ -1557,6 +1797,7 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
 #    - vtype [string]: String indicating whether vfile indicates a byte string ('string') or a file ('file').          #
 #    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
+#    - vwarnings [boolean]: If warnings should be given out or not.                                                    #
 #                                                                                                                      #
 #  Output:                                                                                                             #
 #    - vrecords [list]: A list of lists, containing the results of a hmmer search as lists stored in a list.           #
@@ -1564,10 +1805,7 @@ def f_run_sequences_through_pfam(vhead, vseq, vsavefolder, vjobid, vwarnings_pfa
 ########################################################################################################################
 
 
-def f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid):
-    # Parameters
-    vsilent = 1
-
+def f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid, vwarnings):
     # Open file for real if type is file
     if vtype == 'file':
         # Test if file can be opened, else abort.
@@ -1595,7 +1833,7 @@ def f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid):
     for vl, vline in enumerate(vfh_xml):
         vline_stripped = vline.strip()
         if vline_stripped == '':
-            if vsilent == 0:
+            if vwarnings:
                 print('Warning: Empty line: ' + str(vl))
         else:
             vtypeline = f_check_type_line(vline)
@@ -1604,6 +1842,9 @@ def f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid):
                 vrecords.append(vrecord)
             elif vtypeline == 2:
                 vcurr_parent = f_get_parent(vrecords, vline)
+            elif vtypeline == 3:  # comments
+                if vwarnings:
+                    print('Comment on line ' + str(vl) + ': ' + vline)
             else:
                 print('Line is not a correct xml line.')
                 print(vline)
@@ -1674,6 +1915,8 @@ def f_check_type_line(vline):
         vreturn = 2
     elif vline.startswith('<') and vline.endswith('>'):
         vreturn = 1
+    elif vline.startswith('//'):
+        vreturn = 3  # comment
     else:
         vreturn = -1
     return vreturn
@@ -1758,6 +2001,7 @@ def f_read_single_record(vline, vcurr_parent, vtype):
 #    - vtype [string]: String indicating whether vfile indicates a byte string ('string') or a file ('file').          #
 #    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
+#    - vwarnings [boolean]: If warnings should be given out or not.                                                    #
 #                                                                                                                      #
 #  Output:                                                                                                             #
 #    - vheader [list]: Header indicators with the first few headers being the same as in the tsv output of hmmer.      #
@@ -1766,8 +2010,8 @@ def f_read_single_record(vline, vcurr_parent, vtype):
 ########################################################################################################################
 
 
-def f_convert_hmmer_xml_tsv(vfile, vtype, vsavefolder, vjobid):
-    vrecords = f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid)
+def f_convert_hmmer_xml_tsv(vfile, vtype, vsavefolder, vjobid, vwarnings):
+    vrecords = f_read_hmmer_xml(vfile, vtype, vsavefolder, vjobid, vwarnings)
 
     # Create tsv header and the ids that are listed in the xml
     vheader = []
@@ -1841,8 +2085,10 @@ def f_convert_hmmer_xml_tsv(vfile, vtype, vsavefolder, vjobid):
 #    - vseq [string]: Protein sequence.                                                                                #
 #    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
+#    - vwarnings [boolean]: If warnings should be given out or not.                                                    #
 #                                                                                                                      #
 #  Output:                                                                                                             #
+#    - verrorcode [list]: Error code to check what is going on.                                                        #
 #    - vheader [list]: The header of the records, indicating what is what.                                             #
 #    - vrecords [list]: List of lists with every list being a domain, and every sublist an attribute of the domain     #
 #                       record.                                                                                        #
@@ -1850,12 +2096,9 @@ def f_convert_hmmer_xml_tsv(vfile, vtype, vsavefolder, vjobid):
 ########################################################################################################################
 
 
-def f_query_pfam(vseq, vsavefolder, vjobid, viteration, vwaittime):
-    vmax_iterations = 6
-    if viteration >= vmax_iterations:
-        raise ValueError('PFAM currently down (503 error).')
-    time.sleep(vwaittime)
-
+# noinspection PyBroadException
+def f_query_pfam(vseq, vsavefolder, vjobid, vwarnings):
+    verrorcode = -1
     # Modify sequence
     vseq = ''.join(vseq.split())  # remove spaces within the sequence
     vseq = '>Seq\n' + vseq  # Format sequence for running through hmmer.
@@ -1865,57 +2108,48 @@ def f_query_pfam(vseq, vsavefolder, vjobid, viteration, vwaittime):
     vparameters_encoded = urllib.parse.urlencode(vparameters).encode('utf-8')
     vurllib_request_run = urllib.request.Request('https://www.ebi.ac.uk/Tools/hmmer/search/hmmscan',
                                                  vparameters_encoded)
-    vresult_url = urllib.request.urlopen(vurllib_request_run).geturl()
-    vresult_parameters = {'format': 'xml'}
-    vresult_parameters_encoded = urllib.parse.urlencode(vresult_parameters)
-    vresult_url_with_parameters = vresult_url.replace('results', 'download') + '?' + vresult_parameters_encoded
-    vurllib_request_result = urllib.request.Request(vresult_url_with_parameters)
-
-    # Get results
-    vperform_conversion = True
-    vxml_result = ''
-    try:
-        vxml_result = urllib.request.urlopen(vurllib_request_result).read()
-    except HTTPError as e:
-        vperform_conversion = False
-        if e.code == 500:
-            raise ValueError('No matching Pfam domains were found.')
-        elif e.code == 503:
-            # Currently unavailable
-            viteration += 1
-            if viteration == 1:
-                vwaittime = 5
-            else:
-                vwaittime *= 4
-            try:
-                f_query_pfam(vseq, vsavefolder, vjobid, viteration, vwaittime)
-            except ValueError:
-                f_write_log(vsavefolder, vjobid, 'HTTPError 503 occurred while trying to reach PFAM.\n', 'a')
-                f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError 503 occurred while trying to reach PFAM.')
-                raise ValueError('PFAM currently down (503 error).')
-                sys.exit()
-        elif e.code == 404:
-            f_write_log(vsavefolder, vjobid, 'HTTPError 404 occurred while trying to reach PFAM.\n', 'a')
-            f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError ' + str(e.code) +
-                                                    ' occurred while trying to reach PFAM.')
-            raise ValueError('PFAM currently down (404 error).')
-            sys.exit()
-        else:
-            f_write_log(vsavefolder, vjobid, 'HTTPError ' + str(e.code) + ' occurred while trying to reach PFAM.\n',
-                        'a')
-            print('HTTPError ' + str(e.code) + ' occurred.')
-            print(e.reason)
-            # Write kill cookie
-            f_write_cookie(-1, vsavefolder, vjobid, 'HTTPError ' + str(e.code) + ' occurred while trying to reach' 
-                                                    ' PFAM.')
-            sys.exit()
-
-    # Modify results such that it comes in form of a table
+    vperform_retrieval = True
+    vperform_conversion = False
     vheader = []
     vrecords = []
+    try:
+        vresult_url = urllib.request.urlopen(vurllib_request_run).geturl()
+    except HTTPError as e:
+        vperform_retrieval = False
+        verrorcode = e.code
+        f_write_log(vsavefolder, vjobid, e.reason + '\n', 'a')
+    except:
+        vperform_retrieval = False
+        verrorcode = -2
+
+    if vperform_retrieval:
+        vresult_parameters = {'format': 'xml'}
+        vresult_parameters_encoded = urllib.parse.urlencode(vresult_parameters)
+        # noinspection PyUnboundLocalVariable
+        vresult_url_with_parameters = vresult_url.replace('results', 'download') + '?' + vresult_parameters_encoded
+        vurllib_request_result = urllib.request.Request(vresult_url_with_parameters)
+
+        # Get results
+        vperform_conversion = True
+        vxml_result = ''
+        try:
+            vxml_result = urllib.request.urlopen(vurllib_request_result).read()
+        except HTTPError as e:
+            vperform_conversion = False
+            if e.code == 500:
+                verrorcode = -1  # Is ok, This means, that no domains were found.
+            else:
+                verrorcode = e.code
+                f_write_log(vsavefolder, vjobid, e.reason + '\n', 'a')
+        except:
+            vperform_conversion = False
+            verrorcode = -2
+
+    # Modify results such that it comes in form of a table
     if vperform_conversion:
-        vheader, vrecords = f_convert_hmmer_xml_tsv(vxml_result, 'string', vsavefolder, vjobid)
-    return vheader, vrecords
+        # noinspection PyUnboundLocalVariable
+        vheader, vrecords = f_convert_hmmer_xml_tsv(vxml_result, 'string', vsavefolder, vjobid, vwarnings)
+    return verrorcode, vheader, vrecords
 
 
 ########################################################################################################################
@@ -1930,6 +2164,7 @@ def f_query_pfam(vseq, vsavefolder, vjobid, viteration, vwaittime):
 #    - ventry_found [list]: Is the entry that should be saved. First item is the sequence itself.                      #
 #    - vsavefolder [string]: Indicates the location of the folder where result files should be saved.                  #
 #    - vjobid [string]: Id of the run. Is used to produce the output file names.                                       #
+#    - vcurrentdate [string]: Indicates the currend date that the record is from.                                      #
 #                                                                                                                      #
 #  Output:                                                                                                             #
 #    - no direct output                                                                                                #
@@ -2205,13 +2440,15 @@ def f_dissect_pfam_key(vout, key, value):
 #    - vmaxcutoff_ri [string] gets converted into float: Same as vcutoff, however, before the domain could exist       #
 #                                                        anywhere in the protein. Here the domains need to be present  #
 #                                                        at the same location to make the cut.                         #
-#    - vcustom_scaling_on_ri [string] gets converted into boolean: Allows for custom scaling of the figure. The \      #
+#    - vcustom_scaling_on_ri [string] gets converted into boolean: Allows for custom scaling of the figure. The        #
 #                                                                  standard case is that this is not on, and thus the  #
 #                                                                  scaling is the same regardless of the size of the   #
-#                                                                  protein. If this is on, one can set sbp (default    #
-#                                                                  100 bp per inch) to scale the width of the figure.  #
-#    - vscalingfigure_ri [string] gets converted into float: Indicates the number of bp that are displayed per inch of #
-#                                                            x-axis.                                                   #
+#                                                                  protein. If this is on, one can set with option api #
+#                                                                  (default 100 amino acids per inch) the number of    #
+#                                                                  amino acids displayed per inch to scale the width   #
+#                                                                  of the figure.                                      #
+#    - vscalingfigure_ri [string] gets converted into float: Indicates the number of amino acids per inch that are     #
+#                                                         displayed per inch of x-axis.                                #
 #    - vabsolute_ri [string] gets converted into True/False: Indicates whether absolute numbers are displayed on the   #
 #                                                            y-axis.                                                   #
 #    - vwarnings_ri [string] gets converted into True/False: Indicates whether warnings are written out.               #
@@ -2237,7 +2474,6 @@ def f_reinitialize_dbs(vjobid_ri, vignoredb_ri, vsavefolder_ri, vdbfolder_ri, vg
                        vcolorfile_ri, vignoredomainfile_ri, vcutoff_ri, vmaxcutoff_ri, vcustom_scaling_on_ri,
                        vscalefigure_ri, vabsoluteresults_ri, vwarnings_ri, vfrom_scratch_ri, vnotolderthan_ri,
                        vtemp_db_place_ri):
-
     # Create fasta input file from the databases
     vfastafile = join(vtemp_db_place_ri, vjobid_ri + '_reinitialize.fa')
     vfh_ri_out = open(vfastafile, 'w')
@@ -2261,13 +2497,11 @@ def f_reinitialize_dbs(vjobid_ri, vignoredb_ri, vsavefolder_ri, vdbfolder_ri, vg
             vfh_ri_out.write('\n')
     vfh_ri_out.close()  # Close temporary Fasta file
 
-    # Run propplot
-    f_run_propplot(vjobid_ri, vfastafile, vignoredb_ri, vsavefolder_ri, vtemp_db_place_ri, vgroupfile_ri,
-                   vcolorfile_ri, vignoredomainfile_ri, vcutoff_ri, vmaxcutoff_ri, vcustom_scaling_on_ri,
-                   vscalefigure_ri, vabsoluteresults_ri, vwarnings_ri, vfrom_scratch_ri, vnotolderthan_ri)  # Run the
+    # Run domainviz
+    f_run_domainviz(vjobid_ri, vfastafile, vignoredb_ri, vsavefolder_ri, vtemp_db_place_ri, vgroupfile_ri,
+                    vcolorfile_ri, vignoredomainfile_ri, vcutoff_ri, vmaxcutoff_ri, vcustom_scaling_on_ri,
+                    vscalefigure_ri, vabsoluteresults_ri, vwarnings_ri, vfrom_scratch_ri, vnotolderthan_ri)  # Run the
     # domain search
-
-    print('Still has to be implemented')
 
     # Replace old databases with new ones
     shutil.rmtree(vdbfolder_ri)  # Remove old databases
@@ -2291,16 +2525,16 @@ def f_reinitialize_dbs(vjobid_ri, vignoredb_ri, vsavefolder_ri, vdbfolder_ri, vg
 ########################################################################################################################
 #                                                                                                                      #
 #  f_print_help                                                                                                        #
-#  Prints the help page of propplot.py                                                                                 #
-#  Can be activated by either running just propplot.py or propplot.py -h                                               #
+#  Prints the help page of domainviz.py                                                                                #
+#  Can be activated by either running just domainviz.py or domainviz.py -h                                             #
 #                                                                                                                      #
 ########################################################################################################################
 
 
 def f_print_help():
-    print('Function: propplot.py\n'
+    print('Function: domainviz.py\n'
           'Function description:\n'
-          'propplot is a script that takes protein sequences as an input, runs them through Prosite and PFAM and\n'
+          'domzinviz is a script that takes protein sequences as an input, runs them through Prosite and PFAM and\n'
           'evaluates which predicted protein domains are existing in many of the proteins delivered and displays\n'
           'the results as a histogram, relative to the median of the protein length.\n'
           '\n'
@@ -2327,7 +2561,8 @@ def f_print_help():
           '                                  is 0.5, 50% of the sequences have to have this domain at the same \n'
           '                                  relative position).\n'
           '  -cs [0 or 1]: Can set custom scaling of figure on (default is off = 0).\n'
-          '  -sbp [number larger than 0]: Indicates the scaling of bp per inch that is used for figures.\n'
+          '  -api [number larger than 0]: Indicates the scaling of amino acids per inch of x axis that is used for\n'
+          '                               figures.\n'
           '  -fs [0 or 1]: Indicates if previous results should be overwritten.'
           '\n'
           'Output:\n'
@@ -2339,8 +2574,8 @@ def f_print_help():
           '  - jobid_domain_color_file.txt [file]: Tsv of the domains and colors used in the plots. Can be modified\n'
           '                                        and funneled back into the script to modify colors. See vcolorfile\n'
           '                                        for structure.\n'
-          '  - jobid_proteingroup_prosite.csv [file]: Results of Prosite for protein group.\n'
-          '  - jobid_proteingroup_pfam.csv [file]: Results of PFAM for protein group.\n'
+          '  - jobid_proteingroup_prosite_domain_results_per_aa.tsv [file]: Results of Prosite for protein group.\n'
+          '  - jobid_proteingroup_pfam_domain_results_per_aa.tsv [file]: Results of PFAM for protein group.\n'
           '  - jobid_prosite_res.tsv [file]: Results of Prosite for this job.\n'
           '  - jobid_pfam_res.tsv [file]: Results of PFAM for this job.\n'
           '  - jobid_cookie[id] [file]: Cookie that shows that this part of code was processed. If cookie -1 shows\n'
@@ -2365,11 +2600,12 @@ def f_print_help():
           'Output:\n'
           'New db files at old locations given by parameter -dbf.\n')
 
+
 ########################################################################################################################
 #                                                                                                                      #
 #  Usual entry point of the script.                                                                                    #
-#  propplot is assumed to be started from the command line with arguments (optional and non-optional ones)             #
-#  Please refer to the help (propplot -h) for more information
+#  domainviz is assumed to be started from the command line with arguments (optional and non-optional ones)            #
+#  Please refer to the help (domainviz -h) for more information                                                        #
 #                                                                                                                      #
 ########################################################################################################################
 
@@ -2452,7 +2688,7 @@ if __name__ == "__main__":
         except IOError:
             print('Path ' + vtemp_db_place_main + ' can not be created.')
             sys.exit()
-        if vwarnings_main == 1:  # Give output about where the db location is, if warnings are on.
+        if vwarnings_main == '1':  # Give output about where the db location is, if warnings are on.
             print('Folder to save dbs: ' + vdbfolder_main)
         f_reinitialize_dbs(vjobid_main, vignoredb_main, vsavefolder_main, vdbfolder_main, vgroupfile_main,
                            vcolorfile_main, vignoredomainfile_main, vcutoff_main, vmaxcutoff_main,
@@ -2517,7 +2753,7 @@ if __name__ == "__main__":
                 except IOError:
                     print('Path ' + vsavefolder_main + ' can not be created.')
                     sys.exit()
-                if vwarnings_main == 1:  # Give output about where the storing location is, if warnings are on.
+                if vwarnings_main == '1':  # Give output about where the storing location is, if warnings are on.
                     print('Folder to save intermediate and output files: ' + vsavefolder_main)
             elif sys.argv[vi] == '-dbf':  # Indicates the location of db files.
                 if len(sys.argv) == vi + 1:  # make sure that there is an argument after -dbf.
@@ -2530,7 +2766,7 @@ if __name__ == "__main__":
                 except IOError:
                     print('Path ' + vdbfolder_main + ' can not be created.')
                     sys.exit()
-                if vwarnings_main == 1:  # Give output about where the db location is, if warnings are on.
+                if vwarnings_main == '1':  # Give output about where the db location is, if warnings are on.
                     print('Folder to save dbs: ' + vdbfolder_main)
             elif sys.argv[vi] == '-in':  # Fasta sequence file that contains the protein sequences that should be run.
                 vinputfile_main = sys.argv[vi + 1]
@@ -2541,7 +2777,7 @@ if __name__ == "__main__":
                     print('Can not read ' + vinputfile_main)
                     sys.exit()
                 vfh.close()
-                if vwarnings_main == 1:  # Give output about where the fasta sequences are, if warnings are on.
+                if vwarnings_main == '1':  # Give output about where the fasta sequences are, if warnings are on.
                     print('Input file stored in: ' + vinputfile_main)
             elif sys.argv[vi] == '-gf':  # A file that contains the groupings of proteins.
                 vgroupfile_main = sys.argv[vi + 1]
@@ -2552,7 +2788,7 @@ if __name__ == "__main__":
                     print('Can not read ' + vgroupfile_main)
                     sys.exit()
                 vfh.close()
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Group association information stored in: ' + vgroupfile_main)
             elif sys.argv[vi] == '-cf':  # A file that contains the coloring of domains.
                 vcolorfile_main = sys.argv[vi + 1]
@@ -2563,7 +2799,7 @@ if __name__ == "__main__":
                     print('Can not read ' + vcolorfile_main)
                     sys.exit()
                 vfh.close()
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Domain color information stored in: ' + vcolorfile_main)
             elif sys.argv[vi] == '-if':  # A file that contains domains that should not be printed out.
                 vignoredomainfile_main = sys.argv[vi + 1]
@@ -2574,13 +2810,13 @@ if __name__ == "__main__":
                     print('Can not read ' + vignoredomainfile_main)
                     sys.exit()
                 vfh_main.close()
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Information which domains to ignore stored in: ' + vignoredomainfile_main)
             elif sys.argv[vi] == '-ar':  # Indicates whether absolute or relative results should be used.
                 vabsoluteresults_main = sys.argv[vi + 1]
                 vi += 1
                 if vabsoluteresults_main == '1':
-                    if vwarnings_main == 1:
+                    if vwarnings_main == '1':
                         print('Using absolute results.')
                 elif vabsoluteresults_main != '0':
                     print('Unknown option: ' + sys.argv[vi] + ' ' + sys.argv[vi + 1])
@@ -2588,24 +2824,24 @@ if __name__ == "__main__":
             elif sys.argv[vi] == '-cut':  # Indicates at what cutoff domains should be displayed.
                 vcutoff_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Domain display threshold values: ' + vcutoff_main)
             elif sys.argv[vi] == '-mcut':  # Only domains that have a maximum prevalence at a relative place in the
                 # protein group above this ratio are plotted (e.g. if value is 0.5, 50% of the sequences have to have
                 # this domain at the same relative position).
                 vmaxcutoff_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
-                    print('Domain display threshold values at bp level: ' + vmaxcutoff_main)
-            elif sys.argv[vi] == '-sbp':  # Indicates the scaling of bp per inch.
+                if vwarnings_main == '1':
+                    print('Domain display threshold values at amino acid level: ' + vmaxcutoff_main)
+            elif sys.argv[vi] == '-api':  # Indicates the scaling of amino acids per inch.
                 vscalefigure_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Figure scale is 100pb per ' + vmaxcutoff_main + 'inch.')
-            elif sys.argv[vi] == '-cs':  # Indicates the scaling of bp per inch.
+            elif sys.argv[vi] == '-cs':  # Indicates the scaling of amino acids per inch.
                 vcustom_scaling_on_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     if vcustom_scaling_on_main == '1':
                         print('Custom scaling is on.')
                     else:
@@ -2613,12 +2849,12 @@ if __name__ == "__main__":
             elif sys.argv[vi] == '-fs':
                 vfrom_scratch_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('All previous results ignored.')
             elif sys.argv[vi] == '-not':
                 vnotolderthan_main = sys.argv[vi + 1]
                 vi += 1
-                if vwarnings_main == 1:
+                if vwarnings_main == '1':
                     print('Results recovered from databases should not be older than ' + vnotolderthan_main + '.')
             vi += 1
 
@@ -2655,11 +2891,11 @@ if __name__ == "__main__":
         # Produce output about the scaling of the figure
         if vstandardscalefigure_main == vscalefigure_main:
             if vwarnings_main == '1':
-                print('Using figure scaling of 100 bp per ' + str(vscalefigure_main) + ' inches. Change setting with ' +
-                      'scalefigure or -s followed by a value greater than 0.')
-
+                print('Using standard figure scaling. Change setting with scalefigure or -api followed by a value '
+                      'greater than 0 in combination with -cs 1.')
         # Run the script and produce the plots
         print('Running script:')
-        f_run_propplot(vjobid_main, vinputfile_main, vignoredb_main, vsavefolder_main, vdbfolder_main, vgroupfile_main,
-                       vcolorfile_main, vignoredomainfile_main, vcutoff_main, vmaxcutoff_main, vcustom_scaling_on_main,
-                       vscalefigure_main, vabsoluteresults_main, vwarnings_main, vfrom_scratch_main, vnotolderthan_main)
+        f_run_domainviz(vjobid_main, vinputfile_main, vignoredb_main, vsavefolder_main, vdbfolder_main, vgroupfile_main,
+                        vcolorfile_main, vignoredomainfile_main, vcutoff_main, vmaxcutoff_main, vcustom_scaling_on_main,
+                        vscalefigure_main, vabsoluteresults_main, vwarnings_main, vfrom_scratch_main,
+                        vnotolderthan_main)
